refactor: extract account modal logic into a custom hook for better separation of concerns
